refactor: move parse datetime to utils
